refactor : 숫자에 언더스코어 추가 및 LottoType 네이밍을 RankType으로 변경
